Adding initial version of 60_Registration_Introduction notebook.
Notebook 60_Registration_Introduction uses a shared resource (image),
which is used as a link both in the Python and R versions of the
notebook.

The share resource is in the SharedImages directory and there are
symbolic links to it from the Python and R directories. This allows
the notebook server to correctly display the image no matter where it
is launched.