More R style code
Main differences in code to date are to do with the use of name
attributes that simplify the data structures when reading dicoms.

All data structures in R can have names for components - i.e.
everything can be a dictionary. This means you can avoid having
the first element of the list as the series id by using a list
named by series id.